safari-600.7-branch/LayoutTests/mathml/presentation/inferred-mrow-stretchy-expected.txt

This test passes if all the vertical bars stretch.

 
 
 
 
 
 

Mrow: PASS
Sqrt: PASS
Style: PASS
Error: PASS
Phantom: PASS
Math: PASS
Menclose: PASS
